+++ CumulusMX/Api.cs
@@ -288,7 +288,7 @@ namespace CumulusMX
 							await writer.WriteAsync(Station.GetIntervalData(from, to, data));
 							break;
 						case "dailydata.json":
-							await writer.WriteAsync(Station.GetDailylData(from, to, data));
+							await writer.WriteAsync(Station.GetDailyData(from, to, data));
 							break;
 						default:
 							Response.StatusCode = 404;